requirements.txt
django
djangorestframework
django-cors-headers
django-ninja
uvicorn
psycopg
black